Remove hub-native-op-handler and move the functionality to request-handler
